requirements.txt
openai==0.27.2
python-dotenv==1.0.0
pyttsx3==2.90
speechrecognition==3.10.0
opencv-python==4.7.0.72
